# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoneyExchangeTHGroup Admin API",
    "description": "RESTful API for the MoneyExchangeTHGroup Admin Panel. Manages currencies, exchange rates, members, bookings, wallet top-ups, articles, and admin profiles.\n\nAll endpoints (except `GET /api/currencies` and `GET /api/articles?public=true`) require authentication via Supabase session cookies.",
    "version": "1.0.0",
    "contact": {
      "name": "Nexmind IT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/api/currencies": {
      "get": {
        "tags": ["Currencies"],
        "summary": "List all currencies",
        "description": "Returns all currencies sorted by code. **Public endpoint** — no authentication required.",
        "operationId": "listCurrencies",
        "responses": {
          "200": {
            "description": "Currency list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Currency" }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Currencies"],
        "summary": "Add new currency",
        "description": "Create a new currency entry. Requires authentication.",
        "operationId": "createCurrency",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code", "name"],
                "properties": {
                  "code": { "type": "string", "example": "USD", "description": "ISO 4217 currency code" },
                  "name": { "type": "string", "example": "US Dollar" },
                  "flag": { "type": "string", "example": "🇺🇸" },
                  "buy_rate": { "type": "number", "example": 34.25 },
                  "sell_rate": { "type": "number", "example": 35.10 },
                  "buy_margin_percent": { "type": "number", "example": 1.5 },
                  "sell_margin_percent": { "type": "number", "example": 1.5 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Currency created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Currency" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/members": {
      "get": {
        "tags": ["Members"],
        "summary": "List all members",
        "operationId": "listMembers",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["active", "suspended", "banned"] },
            "description": "Filter by member status"
          }
        ],
        "responses": {
          "200": {
            "description": "Member list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Member" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Members"],
        "summary": "Create new member",
        "operationId": "createMember",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email"],
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "email": { "type": "string", "format": "email", "example": "john@example.com" },
                  "phone": { "type": "string", "example": "+66812345678" },
                  "status": { "type": "string", "enum": ["active", "suspended", "banned"], "default": "active" },
                  "wallet_balance": { "type": "number", "default": 0 },
                  "verified": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Member" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/members/{id}": {
      "get": {
        "tags": ["Members"],
        "summary": "Get single member",
        "operationId": "getMember",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Member details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Member" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "tags": ["Members"],
        "summary": "Update member",
        "description": "Partial update of a member. Accepts both snake_case and camelCase field names.",
        "operationId": "updateMember",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string" },
                  "status": { "type": "string", "enum": ["active", "suspended", "banned"] },
                  "wallet_balance": { "type": "number" },
                  "verified": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Member" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Get current admin profile",
        "description": "Returns the profile of the currently authenticated admin user.",
        "operationId": "getProfile",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Profile data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Profile" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "tags": ["Profile"],
        "summary": "Update admin profile",
        "description": "Partial update of the authenticated admin's profile.",
        "operationId": "updateProfile",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Admin User" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string" },
                  "avatar_url": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Profile" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie set after login. Automatically managed by the browser."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Something went wrong" }
        }
      },
      "Profile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Admin User" },
          "email": { "type": "string", "format": "email", "example": "admin@moneyexchangethgroup.com" },
          "avatar_url": { "type": "string", "format": "uri", "nullable": true },
          "phone": { "type": "string", "nullable": true },
          "role": { "type": "string", "enum": ["admin"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Currency": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "code": { "type": "string", "example": "USD" },
          "name": { "type": "string", "example": "US Dollar" },
          "flag": { "type": "string", "example": "🇺🇸" },
          "buy_rate": { "type": "number", "example": 34.25 },
          "sell_rate": { "type": "number", "example": 35.10 },
          "buy_margin_percent": { "type": "number", "example": 1.5 },
          "sell_margin_percent": { "type": "number", "example": 1.5 },
          "last_updated": { "type": "string", "format": "date-time" }
        }
      },
...
      "Member": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "John Doe" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "suspended", "banned"] },
          "join_date": { "type": "string", "format": "date-time" },
          "wallet_balance": { "type": "number", "example": 10000 },
          "verified": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request — validation error or missing fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized — not logged in or session expired",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Unauthorized" }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Resource not found" }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Internal server error" }
              }
            }
          }
        }
      }
    }
  }
}